+++ packages/nft/src/interfaces/events.ts
@@ -1,4 +1,4 @@
-import { PublicKey, Struct, UInt32, Field, Bool } from "o1js";
+import { PublicKey, Struct, UInt64, Field, Bool } from "o1js";
 import { Storage } from "@silvana-one/storage";
 import { NFTStateStruct, UInt64Option } from "./types.js";
 
@@ -38,7 +38,7 @@ class UpdateEvent extends Struct({
   /** The approved address of the NFT after the update. */
   approved: PublicKey,
   /** The version number of the NFT state. */
-  version: UInt32,
+  version: UInt64,
   /** Indicates whether the NFT is paused after the update. */
   isPaused: Bool,
   /** The hash of the verification key used for metadata proofs. */
